Rename `commitinfo.rs` to snake case
